// apps/api/src/graphql/types.generated.d.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
    ID: {
        input: string;
        output: string;
    };
    String: {
        input: string;
        output: string;
    };
    Boolean: {
        input: boolean;
        output: boolean;
    };
    Int: {
        input: number;
        output: number;
    };
    Float: {
        input: number;
        output: number;
    };
    Date: {
        input: any;
        output: any;
    };
    DateTime: {
        input: any;
        output: any;
    };
    SchoolId: {
        input: any;
        output: any;
    };
};
// ...
export type Class = {
    __typename?: 'Class';
    _count?: Maybe<ClassCount>;
    defaultRoom?: Maybe<Room>;
    group?: Maybe<Group>;
    groupId: Scalars['ID']['output'];
    id: Scalars['ID']['output'];
    level: Scalars['String']['output'];
    name: Scalars['String']['output'];
    section?: Maybe<Scalars['String']['output']>;
    students?: Maybe<Array<Student>>;
    supervisor?: Maybe<Teacher>;
    teachingTeamMembers?: Maybe<Array<TeachingTeamMember>>;
    totalCoefficient?: Maybe<Scalars['Int']['output']>;
    totalWeeklyHours?: Maybe<Scalars['Float']['output']>;
};
// ...
export type GetLessonsInput = {
    classId?: InputMaybe<Scalars['ID']['input']>;
    department?: InputMaybe<Scalars['String']['input']>;
    groupId?: InputMaybe<Scalars['ID']['input']>;
    hasLessonOnly?: InputMaybe<Scalars['Boolean']['input']>;
    level?: InputMaybe<Scalars['String']['input']>;
    limit?: Scalars['Int']['input'];
    mode: ResourceMode;
    page?: Scalars['Int']['input'];
    searchTerm?: InputMaybe<Scalars['String']['input']>;
    section?: InputMaybe<Scalars['String']['input']>;
    status?: InputMaybe<LessonStatus>;
    teacherId?: InputMaybe<Scalars['ID']['input']>;
};
// ...
export declare enum ResourceMode {
    Class = "CLASS",
    Teacher = "TEACHER"
}
// ...
export type Teacher = {
    __typename?: 'Teacher';
    assignments?: Maybe<Array<Maybe<TeacherAssignments>>>;
    attendances?: Maybe<Array<AttendanceRecord>>;
    bio?: Maybe<Scalars['String']['output']>;
    createdAt?: Maybe<Scalars['DateTime']['output']>;
    department?: Maybe<Scalars['String']['output']>;
    diploma?: Maybe<Scalars['String']['output']>;
    experience?: Maybe<Scalars['String']['output']>;
    hireDate?: Maybe<Scalars['DateTime']['output']>;
    id: Scalars['ID']['output'];
    isActive?: Maybe<Scalars['Boolean']['output']>;
    salary?: Maybe<Scalars['Float']['output']>;
    schoolUserId?: Maybe<Scalars['ID']['output']>;
    specialization?: Maybe<Scalars['String']['output']>;
    supervisedClasses?: Maybe<Array<Maybe<Class>>>;
    updatedAt?: Maybe<Scalars['DateTime']['output']>;
    user?: Maybe<User>;
    weeklyHours?: Maybe<Scalars['Float']['output']>;
};